added > oqs-ssh-viz@0.1.0 build > react-scripts build
Creating an optimized production build...
Compiled with warnings.

src/BarPlot.tsx
  Line 1:16:  'useState' is defined but never used  @typescript-eslint/no-unused-vars

Search for the keywords to learn more about each warning.
To ignore, add // eslint-disable-next-line to the line before.

File sizes after gzip:

  86.35 KB  build/static/js/2.1e670eb0.chunk.js
  3.13 KB   build/static/js/main.9b7022fd.chunk.js
  1.57 KB   build/static/js/3.a63bb8a4.chunk.js
  1.17 KB   build/static/js/runtime-main.0b65383d.js
  405 B     build/static/css/main.23797043.chunk.css

The project was built assuming it is hosted at /.
You can control this with the homepage field in your package.json.

The build folder is ready to be deployed.
You may serve it with a static server:

  serve -s build

Find out more about deployment here:

  https://cra.link/deployment to  file in order to improve/clarify installation instructions